gh-1556 add to Customer class and implement DeliveryCustomer
