+++ example/src/index.ts
@@ -1,4 +1,3 @@
-// Order matters!
 export * from "./functions";
 export * from "./variables";
 export * from "./types";
